requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pdfminer.six==20221105
Pillow==9.5.0
pytesseract==0.3.10
nltk==3.8.1
requests==2.31.0  # For handling HTTP requests to the Gemini API
motor==3.2.0  # MongoDB connection
pymongo==4.5.0
python-multipart==0.0.12
python-docx==0.8.11  # For handling Word files